input widgets to separate file
